# improve performance of find-fork-point