<!-- Merging Requirements:
- Please give your PR a title that is release-note friendly
- In order to be merged, you must add the most appropriate category Label (Added, Changed, Fixed) to your PR
-->
<!-- Explain why this is an improvement (Does this add missing functionality, improve performance, or reduce complexity?) -->
This PR sits on top of https://github.com/Chia-Network/chia-blockchain/pull/16457 which should land before this one.

### Purpose:

The current implementation of `find_fork_point_in_chain()` traverses two chains from their peaks backwards to find the most recent common block. As long as the block records it traverses are in the block record cache, that's not a problem. However, when it looks up block records from the database, it needs to parse and instantiate a `BlockRecord` object, just to get to the `prev_hash` field. The header hash of the previous block already has its own column in the database and can be looked up from there much more efficiently.

<!-- Does this PR introduce a breaking change? -->
### Current Behavior:

When traversing blocks not in the cache, `BlockRecord` objects are parsed and instantiated.

### New Behavior:

When traversing blocks not in the cache, the header hash of the previous block is pulled directly from the database. This improves performance. 

### Testing Notes:

`find_fork_point_in_chain()` is exercised by our reorg tests.

### Benchmarks

| test | before | after | after/before |
| --- | --- | --- | --- |
| test_long_reorg | 663.8 s | 538.8 s | 81.2 % |

### profile

before:
![baseline-profile](https://github.com/Chia-Network/chia-blockchain/assets/661450/4fc10952-b650-42a4-991b-eed639f50b96)

The time spend parsing `BlockRecord` is saved.

after:
![optimized](https://github.com/Chia-Network/chia-blockchain/assets/661450/d042c6f6-63f2-459c-8f00-96ca6fb622df)